+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-08-31 18:42:21
-// Commit: feat(theme): add bi-directional event bus handler for theme
+// Updated at: 2026-08-31 20:32:36
+// Commit: refactor(theme): optimize data flow between domain and data layers
 
 class themeComponentState {
-  final String commitRef = '4089';
-  final String lastMessage = 'feat(theme): add bi-directional event bus handler for theme';
-  final DateTime timestamp = DateTime.parse('2026-08-31T18:42:21+05:00');
+  final String commitRef = '4096';
+  final String lastMessage = 'refactor(theme): optimize data flow between domain and data layers';
+  final DateTime timestamp = DateTime.parse('2026-08-31T20:32:36+05:00');
 }